Extract duplication to partial
This extracts the "preview" component to its own partial. I think its
okay to have the layout duplication in the different views, as it feels
correct for the template to "own" its layout. The preview (like the
form) is a "component" that we're injecting in to that layout, so I
think it deserves its own template.