Extract disclosure log to Case Management backend
Extracts Case Management specifics out of our models and services.

This was quite difficult to separate into parts given how coupled
`DisclosureLog` and `PublishedRequest` were through the `Infreemation`
specific attributes.

`app/services/disclosure_log.rb`
--------------------------------

Here we provide a `CaseManagement` to the `DisclosureLog` as the
mechanism for getting `published_requests`. I've converted the
`query_params` to more generic names.

Instead of passing the attributes through to
`PublishedRequest.create_update_or_destroy_from_api!` we now just pass
the request object. The `PublishedRequest` returned through the
`CaseManagement` should implement a common interface to its attributes.

`spec/services/disclosure_log_spec.rb`
--------------------------------------

Use doubles with the correct API instead of relying on
CaseManagement-specific attributes.

Tweaked some comments and spec descriptions for added clarity.

Not much changed here.

`app/models/published_request.rb`
---------------------------------

Removes the callback – and its related private methods – that extract
the CaseManagement-specific attributes from the `payload` attribute.
Instead we rely on the `request` duck type passed to
`PublishedRequest.create_update_or_destroy_from_api!` to implement
`#to_h` that maps the CaseManagement attributes to `PublishedRequest`'s
`ActiveRecord` attributes.

All this extraction from the `payload` is now handled by
`CaseManagement::Infreemation::PublishedRequest`.

One gotcha that we need to handle for Infreemation is that we need to
check whether a request returned by the API is actually publishable. To
do this I've added a virtual accessor so that we can use this as a hook
  to check whether to update or destroy our cached copy.

`spec/models/published_request_spec.rb`
---------------------------------------

As in `DisclosureLog`, we're now just passing a request object through
to `PublishedRequest.create_update_or_destroy_from_api!` and then
calling some methods on it, so reflect that in the specs.

The big change here is removing all the code for extracting information
from the `payload` to our `ActiveRecord` attributes – quite simple – and
relying on the `publishable` virtual attribute for deciding whether to
update or destroy our cached copy. The specs have been updated to
reflect this, rather than using the `datepublished` attribute in the
`payload`.