Allow dynamic configuration of CaseManagements
Adds `.configure!` for `CaseManagement` adapters so that we can
configure the underlying library at runtime.

Ideally we’d have a client instance that gets configured on
initialisation, but this at least paves the way for extracting that out.

The important thing is having a common interface for configuration so
that we can easily switch adapters.

There’s some duplication here of course, but I don’t think it’s work
abstracting since the application isn’t the place to do it – the library
is the place to handle configuration details.