Configurable current CaseManagement
When `CaseManagement.current` is called, we pull the config from
`config/case_management.yml` and dynamically configure the specified
client library.

We then initialise the specified `CaseManagement` adapter and cache it.

Ideally we’d have a single step…

    @current ||= "CaseManagement::#{config[:adapter]}".
      constantize.new(config[:client_params])

…but currently we can’t configure the clients on initialisation, so we
must resort to a separate step of configuration.

Also adds a setup step to create a development configuration file, with
Infreemation as the current default.

Creates a `config/case_management.yml` for CI so that we don't have to
handle edge case of file not existing. The actual configuration doesn't
matter as we use generic doubles for the places that interact with a
`CaseManagement`