Add CaseManagement::Icasework#generate_url
iCasework document URLs contain a time-limited access token, so we must
fetch the latest URL each time we need to present it.

The system generated responses all have a name of "Response (foo bar
baz)", but it is possible that a document with "Response" at the start
of the name is manually uploaded.

The documents are returned in alphabetical order, so we cannot rely on
the first document with a name beginning "Response" to be the correct
document.

Filtering Name on the Regexp and checking for PDFs gives us the best
chance of identifying the response document.