Handle icasework-ruby KeyErrors
LazyHash uses `fetch`, so when a key does not exist we get a KeyError.
It seems that sometimes these keys don't exist on some cases, so we need
to handle them like a standard Hash and return nil if there is no key.

Using `#fetch(key, nil)` in the client library would probably solve
this.